requirements.txt
pandas
numpy
scipy
statsmodels
matplotlib
lark-parser
jinja2
tabulate
